Make it possible to build OpenVINO
This recipe is still relying on the user to download a readily compiled
openvino release to a very absolute and specific location.

This also does not include libtbb.so.* in the wheel for aarch64
and also doesn't seem to set the properties for the wheel correctly so
that openvino is properly packaged so that no `LD_LIBRARY_PATH`
modification is necessary.