Implement `logits_filter_callback`
This is exposed via `on_new_logits` callback in the Python API
to be consistent with the `on_new_...` scheme.

Note that currently only the number of tokens so far and the
raw logits are exposed. So to be truly compatible with whisper.cpp
the other arguments would need to be exposed to Python as well
but it is better than nothing.

The logits are mutable so that the callback can indeed influence the
decoding process. By using the `array_t` type this also safes us a copy.